gereksiz endpointler kaldırıldı
