+++ src/index.ts
@@ -7,13 +7,10 @@ import userRoutes from './routes/userRoutes';
 import authRoutes from './routes/authRoutes';
 import eventRoutes from './routes/eventRoutes';
 import sportsRoutes from './routes/sportsRoutes';
-import securityRoutes from './routes/securityRoutes';
-import reportRoutes from './routes/reportRoutes';
 import profileRoutes from './routes/profileRoutes';
 import statsRoutes from './routes/statsRoutes';
 import statisticsRoutes from './routes/statisticsRoutes';
 import userNewsRoutes from './routes/userNewsRoutes';
-import notificationRoutes from './routes/notificationRoutes';
 import announcementRoutes from './routes/announcementRoutes';
 import eventRatingRoutes from './routes/eventRatingRoutes';
 import userReportRoutes from './routes/userReportRoutes';
@@ -78,11 +75,8 @@ app.use('/api/users', userRoutes);
 app.use('/api/profile', profileRoutes);
 app.use('/api/events', eventRoutes);
 app.use('/api/sports', sportsRoutes);
-app.use('/api/security', securityRoutes);
-app.use('/api/reports', reportRoutes);
 app.use('/api/stats', statsRoutes);
 app.use('/api/user-news', userNewsRoutes);
-app.use('/api/notifications', notificationRoutes);
 app.use('/api/announcements', announcementRoutes);
 app.use('/api/event-ratings', eventRatingRoutes);
 app.use('/api/user-reports', userReportRoutes);
