+++ src/middleware/swaggerMiddleware.ts
@@ -35,15 +35,12 @@ export const setupSwagger = (app: Express): void => {
         { name: 'Auth', description: 'Kimlik doğrulama işlemleri' },
         { name: 'Events', description: 'Etkinlik yönetimi' },
         { name: 'Profile', description: 'Kullanıcı profil yönetimi' },
-        { name: 'Reports', description: 'Rapor yönetimi' },
         { name: 'Users', description: 'User management operations' },
-        { name: 'Security', description: 'Security logs' },
         { name: 'Sports', description: 'Spor yönetimi' },
         { name: 'Stats', description: 'İstatistik ve dashboard verileri' },
-        { name: 'News Scraper', description: 'Haber scraping işlemleri' },
         { name: 'Announcements', description: 'Duyuru yönetimi' },
-        { name: 'Notifications', description: 'Bildirim yönetimi' },
         { name: 'Mobile Notifications', description: 'Mobil bildirim yönetimi' },
+        { name: 'UserNews', description: 'Kullanıcı haberleri işlemleri' },
       ],
       paths: {
         '/api/stats/weekly': {
@@ -140,423 +137,6 @@ export const setupSwagger = (app: Express): void => {
               },
               '401': { $ref: '#/components/responses/UnauthorizedError' },
               '403': { $ref: '#/components/responses/ForbiddenError' },
-              '500': { $ref: '#/components/responses/InternalServerError' }
-            }
-          }
-        },
-        '/api/news-scraper/scraped': {
-          get: {
-            tags: ['News Scraper'],
-            summary: 'Tüm scrape edilmiş haberleri listeler',
-            description: 'Tüm scrape edilmiş haberleri durumlarına bakılmaksızın listeler',
-            security: [{ bearerAuth: [] }],
-            responses: {
-              '200': {
-                description: 'Tüm haberler başarıyla listelendi',
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'object',
-                      properties: {
-                        status: {
-                          type: 'string',
-                          enum: ['success'],
-                          example: 'success'
-                        },
-                        count: {
-                          type: 'number',
-                          example: 10
-                        },
-                        data: {
-                          type: 'array',
-                          items: {
-                            $ref: '#/components/schemas/NewsItem'
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-              },
-              '401': { $ref: '#/components/responses/UnauthorizedError' },
-              '500': { $ref: '#/components/responses/InternalServerError' }
-            }
-          }
-        },
-        '/api/news-scraper/test': {
-          get: {
-            tags: ['News Scraper'],
-            summary: 'API durum kontrolü (Test)',
-            description: "API'nin çalışıp çalışmadığını kontrol etmek için test endpointi (kimlik doğrulaması gerektirmez)",
-            responses: {
-              '200': {
-                description: 'API çalışıyor',
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'object',
-                      properties: {
-                        status: {
-                          type: 'string',
-                          enum: ['success'],
-                          example: 'success'
-                        },
-                        message: {
-                          type: 'string',
-                          example: 'Haber Scraper API çalışıyor!'
-                        },
-                        timestamp: {
-                          type: 'string',
-                          format: 'date-time',
-                          example: '2024-04-27T12:34:56.789Z'
-                        }
-                      }
-                    }
-                  }
-                }
-              }
-            }
-          }
-        },
-        '/api/news-scraper/scrape': {
-          post: {
-            tags: ['News Scraper'],
-            summary: "URL'den haber scrape et",
-            description: "Belirtilen URL'den haber içeriğini scrape eder ve veritabanına kaydeder (admin yetkileri gerektirir)",
-            security: [{ bearerAuth: [] }],
-            requestBody: {
-              required: true,
-              content: {
-                'application/json': {
-                  schema: {
-                    type: 'object',
-                    required: ['url', 'sport_id'],
-                    properties: {
-                      url: {
-                        type: 'string',
-                        description: "Haber scrape edilecek web sitesi URL'si",
-                        example: 'https://www.haberturk.com/spor'
-                      },
-                      sport_id: {
-                        type: 'number',
-                        description: "Spor kategorisi ID'si",
-                        example: 1
-                      }
-                    }
-                  }
-                }
-              }
-            },
-            responses: {
-              '201': {
-                description: 'Haberler başarıyla kaydedildi',
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'object',
-                      properties: {
-                        status: {
-                          type: 'string',
-                          enum: ['success'],
-                          example: 'success'
-                        },
-                        message: {
-                          type: 'string',
-                          example: '5 haber başarıyla kaydedildi'
-                        },
-                        data: {
-                          type: 'array',
-                          items: {
-                            $ref: '#/components/schemas/NewsItem'
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-              },
-              '400': {
-                description: 'Geçersiz istek',
-                content: {
-                  'application/json': {
-                    schema: {
-                      $ref: '#/components/schemas/Error'
-                    },
-                    example: {
-                      status: 'error',
-                      message: "Haber URL'si gereklidir"
-                    }
-                  }
-                }
-              },
-              '404': {
-                description: 'Haber bulunamadı',
-                content: {
-                  'application/json': {
-                    schema: {
-                      $ref: '#/components/schemas/Error'
-                    },
-                    example: {
-                      status: 'error',
-                      message: "URL'den haber bulunamadı"
-                    }
-                  }
-                }
-              },
-              '500': { $ref: '#/components/responses/InternalServerError' }
-            }
-          }
-        },
-        '/api/news-scraper/pending': {
-          get: {
-            tags: ['News Scraper'],
-            summary: 'Bekleyen haberleri listele',
-            description: 'Onay bekleyen haberleri listeler (admin yetkileri gerektirir)',
-            security: [{ bearerAuth: [] }],
-            responses: {
-              '200': {
-                description: 'Bekleyen haberler başarıyla listelendi',
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'object',
-                      properties: {
-                        status: {
-                          type: 'string',
-                          enum: ['success'],
-                          example: 'success'
-                        },
-                        count: {
-                          type: 'number',
-                          example: 3
-                        },
-                        data: {
-                          type: 'array',
-                          items: {
-                            $ref: '#/components/schemas/NewsItem'
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-              },
-              '401': { $ref: '#/components/responses/UnauthorizedError' },
-              '403': { $ref: '#/components/responses/ForbiddenError' },
-              '500': { $ref: '#/components/responses/InternalServerError' }
-            }
-          }
-        },
-        '/api/news-scraper/{id}/status': {
-          patch: {
-             tags: ['News Scraper'],
-             summary: 'Haber durumunu güncelle',
-             description: "Belirtilen ID'ye sahip haberin durumunu güncellers (onayla/reddet)",
-             security: [{ bearerAuth: [] }],
-             parameters: [
-               {
-                 in: 'path',
-                 name: 'id',
-                 required: true,
-                 schema: {
-                   type: 'integer'
-                 },
-                 description: "Durumu güncellenecek haberin ID'si"
-               }
-             ],
-             requestBody: {
-               required: true,
-               content: {
-                 'application/json': {
-                   schema: {
-                     type: 'object',
-                     required: ['status'],
-                     properties: {
-                       status: {
-                         type: 'string',
-                         enum: ['APPROVED', 'REJECTED'],
-                         description: 'Yeni haber durumu'
-                       }
-                     }
-                   }
-                 }
-               }
-             },
-             responses: {
-               '200': {
-                 description: 'Haber durumu başarıyla güncellendi',
-                 content: {
-                   'application/json': {
-                     schema: {
-                       type: 'object',
-                       properties: {
-                         status: { type: 'string', example: 'success' },
-                         message: { type: 'string', example: 'Haber durumu güncellendi' },
-                         data: { $ref: '#/components/schemas/NewsItem' }
-                       }
-                     }
-                   }
-                 }
-               },
-               '400': { description: 'Geçersiz ID veya durum', $ref: '#/components/responses/BadRequestError' },
-               '401': { $ref: '#/components/responses/UnauthorizedError' },
-               '403': { $ref: '#/components/responses/ForbiddenError' },
-               '404': { description: 'Haber bulunamadı', $ref: '#/components/responses/NotFoundError' },
-               '500': { $ref: '#/components/responses/InternalServerError' }
-             }
-           }
-         },
-        '/api/news-scraper/approved': {
-          get: {
-            tags: ['News Scraper'],
-            summary: 'Onaylanmış haberleri listele',
-            description: 'Onaylanmış haberleri listeler',
-            security: [{ bearerAuth: [] }],
-            responses: {
-              '200': {
-                description: 'Onaylanmış haberler başarıyla listelendi',
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'object',
-                      properties: {
-                        status: {
-                          type: 'string',
-                          enum: ['success'],
-                          example: 'success'
-                        },
-                        count: {
-                          type: 'number',
-                          example: 5
-                        },
-                        data: {
-                          type: 'array',
-                          items: {
-                            $ref: '#/components/schemas/NewsItem'
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-              },
-              '401': { $ref: '#/components/responses/UnauthorizedError' },
-              '500': { $ref: '#/components/responses/InternalServerError' }
-            }
-          }
-        },
-        '/api/news-scraper/rejected': {
-          get: {
-            tags: ['News Scraper'],
-            summary: 'Reddedilmiş haberleri listele',
-            description: 'Reddedilmiş haberleri listeler',
-            security: [{ bearerAuth: [] }],
-            responses: {
-              '200': {
-                description: 'Reddedilmiş haberler başarıyla listelendi',
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'object',
-                      properties: {
-                        status: {
-                          type: 'string',
-                          enum: ['success'],
-                          example: 'success'
-                        },
-                        count: {
-                          type: 'number',
-                          example: 3
-                        },
-                        data: {
-                          type: 'array',
-                          items: {
-                            $ref: '#/components/schemas/NewsItem'
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-              },
-              '401': { $ref: '#/components/responses/UnauthorizedError' },
-              '403': { $ref: '#/components/responses/ForbiddenError' },
-              '500': { $ref: '#/components/responses/InternalServerError' }
-            }
-          }
-        },
-        '/api/news-scraper/{id}': {
-          delete: {
-            tags: ['News Scraper'],
-            summary: 'Haberi sil',
-            description: "Belirtilen ID'ye sahip haberi tamamen siler",
-            security: [{ bearerAuth: [] }],
-            parameters: [
-              {
-                in: 'path',
-                name: 'id',
-                required: true,
-                schema: {
-                  type: 'integer'
-                },
-                description: "Silinecek haberin ID'si"
-              }
-            ],
-            responses: {
-              '200': {
-                description: 'Haber başarıyla silindi',
-                content: {
-                  'application/json': {
-                    schema: {
-                      type: 'object',
-                      properties: {
-                        status: { 
-                          type: 'string', 
-                          enum: ['success'],
-                          example: 'success' 
-                        },
-                        message: { 
-                          type: 'string', 
-                          example: 'Haber başarıyla silindi' 
-                        },
-                        data: { 
-                          $ref: '#/components/schemas/NewsItem' 
-                        }
-                      }
-                    }
-                  }
-                }
-              },
-              '400': {
-                description: 'Geçersiz istek',
-                content: {
-                  'application/json': {
-                    schema: {
-                      $ref: '#/components/schemas/Error'
-                    },
-                    example: {
-                      status: 'error',
-                      message: "Geçerli bir haber ID'si gereklidir"
-                    }
-                  }
-                }
-              },
-              '401': { $ref: '#/components/responses/UnauthorizedError' },
-              '403': { $ref: '#/components/responses/ForbiddenError' },
-              '404': {
-                description: 'Haber bulunamadı',
-                content: {
-                  'application/json': {
-                    schema: {
-                      $ref: '#/components/schemas/Error'
-                    },
-                    example: {
-                      status: 'error',
-                      message: 'Haber bulunamadı veya silinemedi'
-                    }
-                  }
-                }
-              },
               '500': { $ref: '#/components/responses/InternalServerError' }
             }
           }
@@ -675,7 +255,6 @@ export const setupSwagger = (app: Express): void => {
                       type: 'object',
                       properties: {
                         status: { type: 'string', example: 'success' },
-                        message: { type: 'string', example: 'Bildirim okundu olarak işaretlendi' },
                         data: { $ref: '#/components/schemas/Notification' }
                       }
                     }
@@ -692,7 +271,7 @@ export const setupSwagger = (app: Express): void => {
           put: {
             tags: ['Notifications'],
             summary: 'Tüm bildirimleri okundu olarak işaretle',
-            description: 'Oturum açmış kullanıcının tüm bildirimlerini okundu olarak işaretler',
+            description: 'Kullanıcının tüm bildirimlerini okundu olarak işaretler',
             security: [{ bearerAuth: [] }],
             responses: {
               '200': {
@@ -703,8 +282,7 @@ export const setupSwagger = (app: Express): void => {
                       type: 'object',
                       properties: {
                         status: { type: 'string', example: 'success' },
-                        message: { type: 'string', example: 'Tüm bildirimler okundu olarak işaretlendi' },
-                        count: { type: 'integer', example: 5 }
+                        affected: { type: 'integer', example: 5 }
                       }
                     }
                   }
@@ -741,7 +319,7 @@ export const setupSwagger = (app: Express): void => {
                       type: 'object',
                       properties: {
                         status: { type: 'string', example: 'success' },
-                        message: { type: 'string', example: 'Bildirim başarıyla silindi' }
+                        message: { type: 'string', example: 'Bildirim silindi' }
                       }
                     }
                   }
@@ -1276,89 +854,6 @@ export const setupSwagger = (app: Express): void => {
               updated_at: { type: 'string', format: 'date-time', description: 'Güncellenme tarihi' }
             }
           },
-          Report: {
-            type: 'object',
-            required: ['id', 'konu', 'raporlayan', 'tarih', 'tur', 'oncelik', 'durum'],
-            properties: {
-              id: {
-                type: 'string',
-                description: 'Benzersiz rapor kimliği'
-              },
-              konu: {
-                type: 'string',
-                description: 'Raporun konusu'
-              },
-              raporlayan: {
-                type: 'string',
-                description: 'Raporu gönderen kullanıcının adı'
-              },
-              tarih: {
-                type: 'string',
-                description: 'Raporun oluşturulma tarihi (GG.AA.YYYY formatında)'
-              },
-              tur: {
-                type: 'string',
-                enum: ['Kullanıcı', 'Etkinlik'],
-                description: 'Raporun türü'
-              },
-              oncelik: {
-                type: 'string',
-                enum: ['Yüksek', 'Orta', 'Düşük'],
-                description: 'Raporun önceliği'
-              },
-              durum: {
-                type: 'string',
-                enum: ['Beklemede', 'İnceleniyor', 'Çözüldü', 'Reddedildi'],
-                description: 'Raporun mevcut durumu'
-              }
-            }
-          },
-          ReportData: {
-            type: 'object',
-            required: ['id', 'subject', 'description', 'reportedBy', 'reportedDate', 'priority', 'status', 'entityId', 'entityType'],
-            properties: {
-              id: {
-                type: 'number',
-                description: 'Benzersiz rapor ID'
-              },
-              subject: {
-                type: 'string',
-                description: 'Raporun konusu'
-              },
-              description: {
-                type: 'string',
-                description: 'Rapor açıklaması'
-              },
-              reportedBy: {
-                type: 'string',
-                description: 'Raporu oluşturan kullanıcı adı'
-              },
-              reportedDate: {
-                type: 'string',
-                format: 'date',
-                description: 'Raporun oluşturulma tarihi (YYYY-MM-DD formatında)'
-              },
-              priority: {
-                type: 'string',
-                enum: ['high', 'medium', 'low'],
-                description: 'Raporun önceliği'
-              },
-              status: {
-                type: 'string',
-                enum: ['pending', 'reviewing', 'resolved', 'rejected'],
-                description: 'Raporun mevcut durumu'
-              },
-              entityId: {
-                type: 'number',
-                description: "Raporlanan varlığın ID'si"
-              },
-              entityType: {
-                type: 'string',
-                enum: ['user', 'event'],
-                description: 'Raporlanan varlığın türü'
-              }
-            }
-          },
           User: {
             type: 'object',
             required: ['id', 'email', 'first_name', 'last_name', 'role'],
@@ -1864,108 +1359,6 @@ export const setupSwagger = (app: Express): void => {
                 },
                 period: { type: 'integer', description: 'Değişimin hesaplandığı gün sayısı', example: 30 }
              }
-          },
-          NotificationType: {
-            type: 'string',
-            enum: [
-              'new_report',
-              'new_event',
-              'event_updated',
-              'user_watched',
-              'user_inactive',
-              'user_active',
-              'new_achievement',
-              'pending_approval',
-              'news_approved',
-              'system_alert'
-            ],
-            description: 'Bildirim türlerini tanımlar'
-          },
-          Notification: {
-            type: 'object',
-            required: ['notification_type', 'content', 'read_status', 'user_id'],
-            properties: {
-              id: {
-                type: 'integer',
-                description: 'Bildirimin benzersiz tanımlayıcısı'
-              },
-              notification_type: {
-                $ref: '#/components/schemas/NotificationType',
-                description: 'Bildirimin türü'
-              },
-              content: {
-                type: 'string',
-                description: 'Bildirim içeriği'
-              },
-              read_status: {
-                type: 'boolean',
-                description: 'Bildirimin okunma durumu'
-              },
-              created_at: {
-                type: 'string',
-                format: 'date-time',
-                description: 'Bildirimin oluşturulma tarihi'
-              },
-              event_id: {
-                type: 'integer',
-                description: 'İlişkili etkinlik ID (varsa)'
-              },
-              user_id: {
-                type: 'string',
-                description: 'Bildirimin hedef kullanıcı ID\'si'
-              },
-              link: {
-                type: 'string',
-                description: 'Bildirimin yönlendireceği link (opsiyonel)'
-              }
-            }
-          },
-          CreateNotificationParams: {
-            type: 'object',
-            required: ['notification_type', 'content', 'user_id'],
-            properties: {
-              notification_type: {
-                $ref: '#/components/schemas/NotificationType',
-                description: 'Bildirimin türü'
-              },
-              content: {
-                type: 'string',
-                description: 'Bildirim içeriği'
-              },
-              event_id: {
-                type: 'integer',
-                description: 'İlişkili etkinlik ID (varsa)'
-              },
-              user_id: {
-                type: 'string',
-                description: 'Bildirimin hedef kullanıcı ID\'si'
-              },
-              link: {
-                type: 'string',
-                description: 'Bildirimin yönlendireceği link (opsiyonel)'
-              }
-            }
-          },
-          ListNotificationsParams: {
-            type: 'object',
-            properties: {
-              user_id: {
-                type: 'string',
-                description: 'Bildirimleri filtrelemek için kullanıcı ID\'si'
-              },
-              read_status: {
-                type: 'boolean',
-                description: 'Bildirim okunma durumuna göre filtreleme'
-              },
-              limit: {
-                type: 'integer',
-                description: 'Sayfalandırma için limit değeri'
-              },
-              offset: {
-                type: 'integer',
-                description: 'Sayfalandırma için başlangıç değeri'
-              }
-            }
           },
           UserDevice: {
             type: 'object',
